+++ ecomerce-shop/src/main/java/com/allan/dto/ProductSummaryDTO.java
@@ -1,15 +1,17 @@
 // src/main/java/com/allan/dto/ProductSummaryDTO.java
 package com.allan.dto;
 
-import lombok.Data;
 import java.util.List;
+
+import lombok.Data;
 
 @Data
 public class ProductSummaryDTO {
     private Long id;
     private String title;
-    private int mrpPrice;
-    private int sellingPrice;
+    // Widened from int to long, matching Product.java.
+    private long mrpPrice;
+    private long sellingPrice;
     private int discountPercent;
     private List<String> images;
     private String color;
