+++ ecomerce-shop/src/main/java/com/allan/model/CartItem.java
@@ -30,9 +30,13 @@ public class CartItem {
 
     private int quantity = 1;
 
-    private Integer mrpPrice;
+    // Widened from Integer to Long: at UGX magnitudes (~3,700x USD), a
+    // moderately priced item at a bulk quantity can exceed Integer's
+    // ~2.1 billion ceiling. Int overflow here would silently wrap to a
+    // garbage or negative value rather than failing loudly.
+    private Long mrpPrice;
 
-    private Integer sellingPrice;
+    private Long sellingPrice;
 
     private Long userId;
